+++ sdk/androidsdk/src/main/java/com/loginradius/androidsdk/activity/WebViewActivity.java
@@ -44,7 +44,7 @@ public class WebViewActivity extends ActionBarActivity {
 
     }
 
-    private void startWebView(String url) {
+   private void startWebView(String url) {
         webView.setWebViewClient(new WebViewClient() {
             public boolean shouldOverrideUrlLoading(WebView view, String url) {
                 if (url.contains("?token")) {
@@ -59,7 +59,11 @@ public class WebViewActivity extends ActionBarActivity {
                         accessToken.apikey = apikey;
                         sendAccessToken(accessToken);
                     }
-                } else {
+                } else if (url.contains("error=access_denied"))
+                {
+                    finish();
+                }else
+                {
                     view.loadUrl(url);
                 }
                 return true;
@@ -69,16 +73,13 @@ public class WebViewActivity extends ActionBarActivity {
         webView.getSettings().setJavaScriptEnabled(true);
         webView.loadUrl(url);
     }
-
+    
     @Override
     // Detect when the back button is pressed
     public void onBackPressed() {
-        if (webView.canGoBack()) {
-            webView.goBack();
-        } else {
-            // Let the system handle the back button
-            super.onBackPressed();
-        }
+        // Let the system handle the back button
+        super.onBackPressed();
+
     }