+++ sdk/androidsdk/src/main/java/com/loginradius/androidsdk/api/PhoneLoginUsingOtpAPI.java
@@ -18,11 +18,12 @@ import retrofit2.HttpException;
 
 public class PhoneLoginUsingOtpAPI {
     public void getResponse(LoginParams value, final AsyncHandler<LoginData> handler) {
+        String smstemplate = (value.getSmsTemplate()!=null) ? value.getSmsTemplate() : "";
         Map<String, String> params = new HashMap<String, String>();
         params.put("apikey", value.getApikey());
         params.put("phone", value.getPhone());
         params.put("otp",value.getOtp());
-        params.put("smstemplate",value.getSmsTemplate());
+        params.put("smstemplate",smstemplate);
 
 
         ApiInterface apiService = RestRequest.getClient().create(ApiInterface.class);